Code xmas list and finalize main page design
Created the xmas list and added all items there as well as decorating
and improving the design on both the landing section and the list
section. Design finalized for now but yet to add links to the list and
to the header anchors. Also made small adjustments to the file
hierarchy of the project.